requirements.txt
numpy>=1.23
pandas>=1.5
scipy>=1.10
scikit-learn>=1.2
matplotlib>=3.7
wfdb>=4.1
pillow>=10.0
opencv-python>=4.8
tqdm>=4.66
torch>=2.1
torchvision>=0.16
